rename experiment to model
