+++ src/easyreflectometry/calculators/bornagain/calculator.py
@@ -3,7 +3,7 @@ __author__ = 'github.com/arm61'
 import numpy as np
 from easyscience.Objects.Inferface import ItemContainer
 
-from easyreflectometry.experiment import Model
+from easyreflectometry.model import Model
 from easyreflectometry.sample import Layer
 from easyreflectometry.sample import Material
 from easyreflectometry.sample import MaterialMixture
