+++ src/easyreflectometry/calculators/refnx/wrapper.py
@@ -5,7 +5,7 @@ from typing import Tuple
 import numpy as np
 from refnx import reflect
 
-from easyreflectometry.experiment.resolution_functions import PercentageFhwm
+from easyreflectometry.model.resolution_functions import PercentageFhwm
 
 from ..wrapper_base import WrapperBase
