+++ src/easyreflectometry/model/model_collection.py
