+++ src/easyreflectometry/model/resolution_functions.py
